Add test/parsearch.c
This quick'n'dirty program is used to check how well our PDB
catalogues perform on Korf's random instances.

+++ search.h
@@ -28,6 +28,6 @@ struct path {
 extern void	path_string(char[PATH_STR_LEN], const struct path *);
 
 /* various */
-extern size_t	search_ida(struct pdb_catalogue *, const struct puzzle *, struct path *, FILE *);
+extern unsigned long long	search_ida(struct pdb_catalogue *, const struct puzzle *, struct path *, FILE *);
 
 #endif /* SEARCH_H */
